add: image選択、プレビュー、削除をNewとEditに移動
